feat: 일정 등록 페이지 메인 기능 구현
