+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import Current from './pages/Current';
+import Current from './pages/current/Current';
 
 function App() {
   return (
