--- src/components/commons/bottomButton/BottomButton.tsx
@@ -1,8 +1,0 @@
-import { Wrapper } from './BottomButton.styles';
-import { bottomButton } from '../../../types/commons/bottomButton';
-
-const BottomButton = ({ text }: bottomButton) => {
-  return <Wrapper>{text}</Wrapper>;
-};
-
-export default BottomButton;
