+++ src/types/bottomButton.ts
@@ -1,3 +1,4 @@
 export interface bottomButton {
   text: string;
+  isActivated: boolean;
 }
